Fix to the naming convention for wrapper classes
Originally, we had assumed that the JSON would have CamelCase class name, i.e.

{'Metadata": {...}}

But this does not match our standard for the data format.
We could change the JSON to match this.
Instead, I am changing the code so that the module name is the key from JSON
while the class name is the module name with the first letter capitalized.

Also add some (commented out) logging to help debug this.