Change the usercache pipeline to work on a per-user basis
Recall that the usercache is user specific. Since the data is all stored in the
same collection right now (the user separation is logical), we could just copy
data from all users over. However, that ties the pipeline to the current
implementation and may not work well with other implementations.

So we switch to a per-user pipeline state, which means that we have to pass the
user_id through everything and modify the tests accordingly.

In addition, we add a method that returns the set of unique user_ids in the
usercache.