Support nullable fields in the place object
The first and last place object in a chain are only partially filled - the
first doesn't have a start and the last doesn't have an end. Without this
change, trying to access the unfilled objects, even to print them, would have
thrown an exception. With this change, we return None instead.