Fix entry wrapper and add support for the statemachine/transition key
Also enhance the handler test to ensure that no entries are stored into the error database.
In our case, the entry was moved into to the long-term database, but the error
occurred after that, so it was also stored into the error database.

So we need to check both the long-term and the error databases.

The test now passes.