Make saving to the usercache more performant
The code to save data to the usercache is now more performant. Before this, we
did not have an index on the usercache. The assumption was that since the
usercache would be cleared out every hour or so, we don't need an index. But as
the load on the usercache increased, adding new values to it got slower and
slower until it started timing out the connection. Which meant that the clients
retried and flooded the usercache with even more data.

Right now, the usercache has a bunch of pending messages, and I am pretty sure
that this is the reason, although I need to verify.

At any rate, creating an index doesn't hurt, and will fix it if/when this falls
apart again. With the index, the timeouts have stopped and the data is again
flowing in smoothly.