Revert WORM -> RO change
In 5e6f8aff239eac4662a8d2a0f53c28638b9b39aa, we changed the location wrappers
from read-only in order to support the creation of a dummy location to
represent the extrapolated first point of the trip. However, in
771189c13f4302895f8fe91496579268841d80c1, we refactored the cleaning code to
use dataframes throughout. So we don't really need this anymore.

Let's go back to the read-only entries and ensure that we don't inadvertently
overwrite location entries.