Significantly simplify and trim down the user-client nexus
we no longer anticipate clients with pre-registration etc, since any such
client would just want to deploy a new UI. This also means that we don't need
to track when the study is active, etc (they can just send a notification that
the study has ended)

this allows us to enormously simplify the client code, and potentially remove
it altogether. The client is just currently a vehicle for storing some vaguely
defined settings. I'm just keeping the code around because it has unit tests -
it it turns out that it is not used, I will remove it.

The new assumption is that the client will be a simple field in the profile to
allow us to dump a particular study's data and send it to them.