Ensure no logging as part of a standard import
Ensure that the database code does not do any logging during module load.  This
is because it is frequently imported by other scripts, and those imports run
before any configuration (in `__main__`) for example is executed. This means
that the logging is configured to the default values, which, given the first
log message, is WARNING.

This change replaces the log with print.
BEFORE: logs only WARNINGS
AFTER: logs everything

+ bonus fix: mark the `-s` parameter as `type=int`. Otherwise it remains a
string and the batching fails later.