Add a script that will load the filled trajectories to the specified server
This is an adapation of the original `load_to_remote_server` script that is
even nicer now that we have restructured the code properly. No weird
dependencies on config code. No copy-pasted tour config.

Just reads the single file (for a user who has exported their own data) or a
timeline (for filled trajectories from synthetic data) and uploads to a server
with "skip" authentication configured.